py_ShowMe/requirements.txt
Django==5.1.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
tzdata==2025.2
sqlparse==0.5.3
asgiref==3.8.1
certifi==2025.1.31
pillow==10.4.0